poo/requirements.txt
Django==2.0.1
psycopg2==2.6.2
ipython-genutils==0.2.0
ipython==6.2.1
pillow==5.1.0
djangorestframework
